backend/requirements.txt
pandas
sqlalchemy
pymongo
python-dotenv
openai
uvicorn